Fix streamXmlElements
The parameter elementName was being shadowed by a local binding in the
match pattern. Because the pattern was simply checking that there was a
second tuple element instead of comparing its value to elementName the
pattern always matched. As a result, the function could not distinguish
between the header and character nodes in the kanjidic2.xml file.